backend/requirements.txt
flask==2.0.1
flask-socketio==5.1.1
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
paho-mqtt==1.6.1
python-dotenv==0.19.0
SQLAlchemy==1.4.23
pymysql==1.0.2
numpy>=1.22.0
pandas>=1.4.0
bcrypt>=4.0.1 